requirements.txt
termcolor==1.1.0
printy==2.2.0
colorama==0.4.4
pycryptodomex==3.10.1
